The first testable iteration of the post
Overview:

Post image deployment configuration is now based on a class hierarchy.
The intent is to build on base classes in the hopes of reusing as much
code as possible.

Implemented in this commit:

Linux
    RedHat
        Enterprise Linux
            Enterprise Linux version 7

In addition to these features, a test harness was applied to crude entry
which selects the bottom most class when the rhel7 target is used.

TODO:
    Target registration
    Target probe functions
    Plugin and override features
    RedHat network configurators
        nmcli, sysconfig/network-scripts, etc
    Documentaion

For more information on the initial thought design, see
post/targets/readme.md