'dynamic' class registration
The need to have things like __meta__ = metaclass(xyz) baked in as a class variable is becoming apparent. I am considering adding a feature to python3 that would give read only access to the assumed __mro__ at import time. For now, this seems to 'work' though I consider this current implementation (down to the target classes) a temporary solution. The current framework is very easy to use, however, but maybe a little thin.